+++ crates/icydb-core/src/types/ulid/mod.rs
@@ -253,7 +253,7 @@ impl Visitable for Ulid {}
 ///
 
 #[cfg(test)]
-mod test {
+mod tests {
     use super::*;
 
     #[test]