+++ crates/icydb/src/db/mod.rs
@@ -3,4 +3,5 @@ pub mod response;
 mod session;
 
 pub use icydb_core::db::response::Row;
+pub use response::{Response, WriteBatchResponse, WriteResponse};
 pub use session::{DbSession, SessionDeleteQuery, SessionLoadQuery};